medicine-dispenser/requirements.txt
boto3==1.34.0
Pillow==10.1.0
pydantic==2.5.3
anthropic==0.7.8
python-dotenv==1.0.0
Flask==3.0.0
gunicorn==21.2.0